% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdi.r
\name{add.alpha}
\alias{add.alpha}
\title{Simple function for adding transparency to colour values, used in plotting by imghist().}
\usage{
add.alpha(col, alpha = 0.75)
}
\arguments{
\item{col}{Colour that transparency should be added to.}

\item{alpha}{Desired opacity.}
}
\value{
Colour hex code with transparency, based on the specified input values.
}
\description{
Simple function for adding transparency to colour values, used in plotting by imghist().
}
\examples{
add.alpha("black",0.5)
}
