% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdi.r
\name{imghist}
\alias{imghist}
\title{Simple histogram plotter for all colour values in an input image. Can be used to help assess whether a chosen threshold value is appropriate.}
\usage{
imghist(image_file, threshold = 0.5, channel = 4)
}
\arguments{
\item{image_file}{Image to be read. Images can be jpeg or png files, or a previously read image saved as an object in R.}

\item{threshold}{Reference value for colour criterium after which pixels that are part of the silhouette should be differentiated from the background.}

\item{channel}{Colour channel to which to apply the threshold criterium. Default is 4 (alpha channel of rgba image). Channel setting needs to be adjusted depending on the colour mode of the image used (e.g. there are two channels to choose from in a greyscale image, and 3 in an rgb image).}
}
\description{
Simple histogram plotter for all colour values in an input image. Can be used to help assess whether a chosen threshold value is appropriate.
}
\examples{
fdir <- system.file(package="gdi")
imghist(file.path(fdir,"exdata","lat.png"))
}
