% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdi.r
\name{rotI}
\alias{rotI}
\title{Calculates the rotational inertia of the body around the y axis. Only works with simple circular/elliptical and rectangular cross-sections, thus pixel-precise estimates are recommended.}
\usage{
rotI(
  x,
  dors_diam = NULL,
  axis_coord = NULL,
  volumes = NULL,
  densities = NULL,
  scale = 1
)
}
\arguments{
\item{x}{Either a data frame that is structured like output of gdi(..., return="all") containing masses and diameters for pixe-wide segments, or a numeric vector of horizontal segment COM positions.}

\item{dors_diam}{An optional vector of transverse diameters of the silhouette, required if not contained in the data.frame supplied to x.}

\item{axis_coord}{An optional x coordinate of the axis of rotation (parallel to the y axis), defaults to the center of mass of the entire volume if not set.}

\item{volumes}{An optional separate vector of volumes, required if x is not a data.frame containing them.}

\item{densities}{An optional vector of segment densities, with length equal to the length or nrow() of x, to be multiplied with the volumes for the COM calculation. If both subtract and densities are supplied, the density is applied only to the "residual" volume that is left after subtraction.}

\item{scale}{Scale value, i.e. number of pixels representing 1 m}
}
\value{
A numeric vector containing four elements: The total mass (on the basis of the gdi results and densities), the rotational inertia of the shape using a point mass approximation of each segment, rotational inertia using a cylindrical approximation for each segment, and rotational inertia using a cuboidal approximation (note that this only changes the mass distribution, while segment masses are still assumed to correspond to gdi results multiplied by optional densities).
}
\description{
Calculates the rotational inertia of the body around the y axis. Only works with simple circular/elliptical and rectangular cross-sections, thus pixel-precise estimates are recommended.
}
\examples{
fdir <- system.file(package="gdi")
measuresil(file.path(fdir,"exdata","lat.png"), return="all")->lat_
measuresil(file.path(fdir,"exdata","dors.png"), return="all")->dors_
gdi(lat_, dors_, return="all")->gdiresults
rotI(gdiresults)
}
