% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdi.r
\name{sellipse.coo}
\alias{sellipse.coo}
\title{calculate coordinates for plotting a superellipse for visualizing body cross-sections}
\usage{
sellipse.coo(k, res = 100)
}
\arguments{
\item{k}{superellipse exponent.}

\item{res}{the desired resolution}
}
\value{
a data frame containing
}
\description{
calculate coordinates for plotting a superellipse for visualizing body cross-sections
}
\examples{
sellipse.coo(2.0)->df #get coordinates for normal ellipse (exponent k=2)
plot(df$x,df$y,col="black", type="l") #plot normal ellipse
sellipse.coo(2.3)->df2 # get coordinates for superellipse with exponent 2.3
lines(df$x,df$y, col="blue") #plot superellipse
}
