% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdi.r
\name{sellipse}
\alias{sellipse}
\title{Estimate area of a superellipse. Assistant function for gdi.}
\usage{
sellipse(a, b, k)
}
\arguments{
\item{a}{First radius of the superellipse.}

\item{b}{Second radius of the superellipse.}

\item{k}{superellipse exponent.}
}
\value{
A single number giving the area of the superellipse
}
\description{
Estimate area of a superellipse. Assistant function for gdi.
}
\examples{
major_radius<-2
minor_radius<-3
exponent<-2.3
sellipse(major_radius, minor_radius, exponent)
}
